HH-218702 move some classes from nab-jdbc-common to nab-jdbc module
