+++ nab-data-source/src/main/java/ru/hh/nab/datasource/NabDataSourceProdConfig.java
@@ -7,8 +7,8 @@ import org.springframework.context.annotation.Configuration;
 import static ru.hh.nab.common.qualifier.NamedQualifier.SERVICE_NAME;
 import ru.hh.nab.datasource.healthcheck.HealthCheckHikariDataSourceFactory;
 import ru.hh.nab.datasource.monitoring.NabMetricsTrackerFactoryProvider;
-import ru.hh.nab.jdbc.common.DatabaseSwitcher;
 import ru.hh.nab.jdbc.common.ext.OpenTelemetryJdbcExtension;
+import ru.hh.nab.jdbc.routing.DatabaseSwitcher;
 import ru.hh.nab.metrics.StatsDSender;
 
 @Configuration
