+++ nab-data-source/src/main/java/ru/hh/nab/datasource/healthcheck/HealthCheckResultImpl.java
@@ -1,6 +1,6 @@
 package ru.hh.nab.datasource.healthcheck;
 
-import ru.hh.nab.jdbc.common.healthcheck.HealthCheckResult;
+import ru.hh.nab.jdbc.healthcheck.HealthCheckResult;
 
 public record HealthCheckResultImpl(boolean healthy) implements HealthCheckResult {
 }
