+++ nab-data-source/src/test/java/ru/hh/nab/datasource/DataSourceContextUnsafeTest.java
@@ -5,9 +5,9 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.slf4j.MDC;
-import static ru.hh.nab.jdbc.common.DataSourceType.MASTER;
-import static ru.hh.nab.jdbc.common.DataSourceType.READONLY;
-import static ru.hh.nab.jdbc.common.DataSourceType.SLOW;
+import static ru.hh.nab.jdbc.DataSourceType.MASTER;
+import static ru.hh.nab.jdbc.DataSourceType.READONLY;
+import static ru.hh.nab.jdbc.DataSourceType.SLOW;
 import static ru.hh.nab.jdbc.routing.DataSourceContextUnsafe.MDC_KEY;
 import static ru.hh.nab.jdbc.routing.DataSourceContextUnsafe.clearMDC;
 import static ru.hh.nab.jdbc.routing.DataSourceContextUnsafe.executeInScope;
