+++ nab-hibernate-persistence-provider/src/test/java/ru/hh/nab/hibernate/NabSessionFactoryBuilderFactoryTest.java
@@ -32,9 +32,9 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import static ru.hh.nab.common.qualifier.NamedQualifier.SERVICE_NAME;
 import ru.hh.nab.hibernate.properties.HibernatePropertiesProvider;
+import ru.hh.nab.jdbc.DataSourceType;
 import ru.hh.nab.jdbc.annotation.ExecuteOnDataSource;
 import ru.hh.nab.jdbc.common.DataSourcePropertiesStorage;
-import ru.hh.nab.jdbc.common.DataSourceType;
 import ru.hh.nab.jdbc.transaction.TransactionalScope;
 import ru.hh.nab.metrics.StatsDSender;
