+++ nab-jdbc/src/main/java/ru/hh/nab/jdbc/healthcheck/HealthCheckResult.java
@@ -1,4 +1,4 @@
-package ru.hh.nab.jdbc.common.healthcheck;
+package ru.hh.nab.jdbc.healthcheck;
 
 public interface HealthCheckResult {
   boolean healthy();
