+++ nab-jdbc/src/main/java/ru/hh/nab/jdbc/routing/DataSourceContext.java
@@ -2,7 +2,7 @@ package ru.hh.nab.jdbc.routing;
 
 import java.util.function.Supplier;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
-import ru.hh.nab.jdbc.common.DataSourceType;
+import ru.hh.nab.jdbc.DataSourceType;
 
 public class DataSourceContext {
   private static boolean checkTransaction = true;
