+++ nab-jdbc/src/main/java/ru/hh/nab/jdbc/routing/DatabaseSwitcher.java
@@ -1,4 +1,4 @@
-package ru.hh.nab.jdbc.common;
+package ru.hh.nab.jdbc.routing;
 
 public interface DatabaseSwitcher {
   String createDataSourceName(String databaseName, String dataSourceType);
