+++ nab-jdbc/src/main/java/ru/hh/nab/jdbc/routing/RoutingDataSource.java
@@ -12,12 +12,12 @@ import org.slf4j.LoggerFactory;
 import org.springframework.jdbc.datasource.DelegatingDataSource;
 import org.springframework.jdbc.datasource.lookup.AbstractRoutingDataSource;
 import org.springframework.lang.Nullable;
+import ru.hh.nab.jdbc.DataSourceType;
 import ru.hh.nab.jdbc.common.DataSourcePropertiesStorage;
-import ru.hh.nab.jdbc.common.DataSourceType;
 import ru.hh.nab.jdbc.common.datasource.NamedDataSource;
 import ru.hh.nab.jdbc.common.ext.JdbcExtension;
-import ru.hh.nab.jdbc.common.healthcheck.HealthCheck;
-import ru.hh.nab.jdbc.common.healthcheck.HealthCheckDataSource;
+import ru.hh.nab.jdbc.healthcheck.HealthCheck;
+import ru.hh.nab.jdbc.healthcheck.HealthCheckDataSource;
 import ru.hh.nab.metrics.Counters;
 import ru.hh.nab.metrics.StatsDSender;
 import ru.hh.nab.metrics.Tag;
