+++ nab-jdbc/src/test/java/ru/hh/nab/jdbc/routing/DataSourceRoutingFilterTest.java
@@ -11,7 +11,7 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import ru.hh.nab.jdbc.common.DataSourceType;
+import ru.hh.nab.jdbc.DataSourceType;
 
 public class DataSourceRoutingFilterTest {
