+++ nab-tests/src/test/java/ru/hh/nab/datasource/DataSourceContextTransactionManagerTest.java
@@ -15,8 +15,8 @@ import org.springframework.transaction.support.DefaultTransactionDefinition;
 import static org.springframework.transaction.support.TransactionSynchronizationManager.isActualTransactionActive;
 import static org.springframework.transaction.support.TransactionSynchronizationManager.isCurrentTransactionReadOnly;
 import static org.springframework.transaction.support.TransactionSynchronizationManager.isSynchronizationActive;
+import ru.hh.nab.jdbc.DataSourceType;
 import ru.hh.nab.jdbc.common.DataSourcePropertiesStorage;
-import ru.hh.nab.jdbc.common.DataSourceType;
 import ru.hh.nab.jdbc.routing.DataSourceContextUnsafe;
 import ru.hh.nab.jpa.JpaTestConfig;
 import ru.hh.nab.jpa.model.TestEntity;
