+++ nab-tests/src/test/java/ru/hh/nab/datasource/ExecuteOnDataSourceAspectTest.java
@@ -26,12 +26,12 @@ import org.springframework.test.context.ContextConfiguration;
 import static org.springframework.transaction.support.TransactionSynchronizationManager.isActualTransactionActive;
 import static org.springframework.transaction.support.TransactionSynchronizationManager.isSynchronizationActive;
 import ru.hh.nab.common.properties.FileSettings;
+import static ru.hh.nab.jdbc.DataSourceType.MASTER;
+import static ru.hh.nab.jdbc.DataSourceType.READONLY;
 import ru.hh.nab.jdbc.annotation.DataSourceCacheMode;
 import ru.hh.nab.jdbc.annotation.ExecuteOnDataSource;
 import ru.hh.nab.jdbc.aspect.ExecuteOnDataSourceAspect;
 import ru.hh.nab.jdbc.aspect.ExecuteOnDataSourceTransactionCallbackFactory;
-import static ru.hh.nab.jdbc.common.DataSourceType.MASTER;
-import static ru.hh.nab.jdbc.common.DataSourceType.READONLY;
 import static ru.hh.nab.jdbc.routing.DataSourceContextUnsafe.getDataSourceName;
 import ru.hh.nab.jpa.JpaTestConfig;
 import ru.hh.nab.testbase.jpa.JpaTestBase;
