+++ nab-tests/src/test/java/ru/hh/nab/jpa/DataSourceSwitchingTest.java
@@ -45,12 +45,12 @@ import ru.hh.nab.datasource.healthcheck.HealthCheckHikariDataSourceFactory;
 import ru.hh.nab.datasource.healthcheck.HealthCheckResultImpl;
 import ru.hh.nab.datasource.healthcheck.UnhealthyDataSourceException;
 import ru.hh.nab.datasource.monitoring.NabMetricsTrackerFactoryProvider;
+import ru.hh.nab.jdbc.DataSourceType;
 import ru.hh.nab.jdbc.common.DataSourcePropertiesStorage;
-import ru.hh.nab.jdbc.common.DataSourceType;
-import ru.hh.nab.jdbc.common.DatabaseSwitcher;
-import ru.hh.nab.jdbc.common.healthcheck.HealthCheckResult;
+import ru.hh.nab.jdbc.healthcheck.HealthCheckResult;
 import static ru.hh.nab.jdbc.routing.DataSourceContext.onDataSource;
 import ru.hh.nab.jdbc.routing.DataSourceContextUnsafe;
+import ru.hh.nab.jdbc.routing.DatabaseSwitcher;
 import ru.hh.nab.jdbc.routing.RoutingDataSource;
 import ru.hh.nab.jdbc.routing.RoutingDataSourceFactory;
 import ru.hh.nab.jpa.model.TestEntity;
