+++ nab-tests/src/test/java/ru/hh/nab/jpa/transaction/DataSourceContextTest.java
@@ -8,9 +8,9 @@ import org.junit.jupiter.api.Test;
 import org.slf4j.MDC;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.transaction.support.TransactionTemplate;
-import static ru.hh.nab.jdbc.common.DataSourceType.MASTER;
-import static ru.hh.nab.jdbc.common.DataSourceType.READONLY;
-import static ru.hh.nab.jdbc.common.DataSourceType.SLOW;
+import static ru.hh.nab.jdbc.DataSourceType.MASTER;
+import static ru.hh.nab.jdbc.DataSourceType.READONLY;
+import static ru.hh.nab.jdbc.DataSourceType.SLOW;
 import static ru.hh.nab.jdbc.routing.DataSourceContext.onDataSource;
 import static ru.hh.nab.jdbc.routing.DataSourceContext.onReplica;
 import static ru.hh.nab.jdbc.routing.DataSourceContext.onSlowReplica;
